Fortune/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.2
django-js-asset==2.0.0
django-mptt==0.14.0
gunicorn==20.1.0
mysql==0.0.3
mysqlclient==2.1.1
sqlparse==0.4.3
urllib3==1.26.15
whitenoise==6.4.0
